test: split unit tests into per-module files
Extract binary-sensor, coordinator and generic-sensor tests out of
test_blaulichtsms.py into dedicated test_binary_sensor.py,
test_coordinator.py and test_sensor.py modules, and move the shared
make_alarm/wrap helpers into _test_fixtures.py. test_blaulichtsms.py
keeps the live-API integration cases gated by SKIP_INTEGRATION_TEST.

Update CLAUDE.md to document the new test layout and invocation.
